Rename custom theme file
