+++ src/app/Providers.tsx
@@ -5,7 +5,7 @@ import { ChakraProvider } from '@chakra-ui/react'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { type FC, type ReactNode } from 'react'
 
-import customTheme from '~/customTheme'
+import customTheme from '~/custom-theme'
 import { queryClient } from '~/query-client'
 
 const Providers: FC<{ children: ReactNode }> = ({ children }) => {
